Tab3 ongoing 2

+++ app/src/main/java/com/example/recyclerview/TravelRegion.java
@@ -8,11 +8,11 @@ public class TravelRegion {
     private String continent;
     private ArrayList<String> type;
     private String money;
-    private Long latitude;
-    private Long longitude;
+    private Double latitude;
+    private Double longitude;
 
     public TravelRegion(String engName, String korName, String continent, ArrayList<String> type, String money,
-                       Long latitude, Long longitude) {
+                       Double latitude, Double longitude) {
         this.engName = engName;
         this.korName = korName;
         this.continent = continent;
@@ -27,14 +27,14 @@ public class TravelRegion {
     public String getContinent() { return this.continent; }
     public ArrayList<String> getType() { return this.type; }
     public String getMoney() { return this.money; }
-    public Long getLatitude() { return this.latitude; }
-    public Long getLongitude() { return this.longitude; }
+    public Double getLatitude() { return this.latitude; }
+    public Double getLongitude() { return this.longitude; }
 
     public void setEngName(String string) { this.engName = string; }
     public void setKorName(String string) { this.korName = string; }
     public void setContinent(String string) { this.continent = string; }
     public void setType(ArrayList<String> arrayList) { this.type = arrayList; }
     public void setMoney(String string) { this.money = string; }
-    public void setLatitude(Long latitude) { this.latitude = latitude; }
-    public void setLongitude(Long longitude) { this.longitude = longitude; }
+    public void setLatitude(Double latitude) { this.latitude = latitude; }
+    public void setLongitude(Double longitude) { this.longitude = longitude; }
 }
